refactor: 重构 Program.cs — DealErrorConfig 数据驱动化 + 提取 UpdateChecker
- DealErrorConfig: 重复if语句 → 元组数组+循环
- CheckUpdate/CheckVersion → Helper/UpdateChecker.cs
- 删除 Main() 中~80行注释代码和废弃方法
- 更新 FmMain.Tray.cs、FmSetting.General.cs 中的调用点
- Program.cs: 900行 → 363行
