+++ TrOCR.csproj
@@ -446,6 +446,7 @@
     <Compile Include="Messageload.Designer.cs">
       <DependentUpon>Messageload.cs</DependentUpon>
     </Compile>
+    <Compile Include="Helper\UpdateChecker.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
